--- app/src/main/java/com/pozmaxpav/cinemaopinion/data/api/models/modelfirebase/FilmData.kt
@@ -1,5 +1,0 @@
-package com.pozmaxpav.cinemaopinion.data.api.models.modelfirebase
-
-data class FilmData(
-    val titleFilm: String = ""
-)
