--- app/src/main/java/com/pozmaxpav/cinemaopinion/data/mappers/Mapperfirebase.kt
@@ -1,16 +1,0 @@
-package com.pozmaxpav.cinemaopinion.data.mappers
-
-import com.pozmaxpav.cinemaopinion.data.api.models.modelfirebase.FilmData
-import com.pozmaxpav.cinemaopinion.domain.models.modelfirebase.FilmDomain
-
-fun FilmData.toDomain(): FilmDomain {
-    return FilmDomain(
-        titleFilm = titleFilm
-    )
-}
-
-fun FilmDomain.toData(): FilmData {
-    return FilmData(
-        titleFilm = titleFilm
-    )
-}
