+++ app/src/main/java/com/pozmaxpav/cinemaopinion/data/repository/repositoryfirebase/FirebaseRepositoryImpl.kt
@@ -2,11 +2,7 @@ package com.pozmaxpav.cinemaopinion.data.repository.repositoryfirebase
 
 import android.util.Log
 import com.google.firebase.database.DatabaseReference
-import com.pozmaxpav.cinemaopinion.data.api.models.modelfirebase.FilmData
-import com.pozmaxpav.cinemaopinion.data.mappers.toData
-import com.pozmaxpav.cinemaopinion.data.mappers.toDomain
 import com.pozmaxpav.cinemaopinion.domain.models.SelectedMovie
-import com.pozmaxpav.cinemaopinion.domain.models.modelfirebase.FilmDomain
 import com.pozmaxpav.cinemaopinion.domain.repository.repositoryfirebase.FirebaseRepository
 import kotlinx.coroutines.tasks.await
 import javax.inject.Inject
@@ -14,14 +10,6 @@ import javax.inject.Inject
 class FirebaseRepositoryImpl @Inject constructor(
     private val databaseReference: DatabaseReference
 ) : FirebaseRepository {
-
-    //    override suspend fun saveMovie(title: String) {
-//        val filmData = FilmDomain(titleFilm = title).toData() // Преобразуем FilmDomain в FilmData
-//        val key = databaseReference.child("list_movies").push().key
-//        key?.let {
-//            databaseReference.child("list_movies").child(it).setValue(filmData).await()
-//        } ?: throw Exception("Failed to generate key")
-//    }
 
     override suspend fun saveMovie(selectedMovie: SelectedMovie) {
         val filmData = SelectedMovie(
@@ -36,7 +24,6 @@ class FirebaseRepositoryImpl @Inject constructor(
     }
 
     override suspend fun removeMovie(id: Double) {
-//        val idToRemove = id.trim().toDoubleOrNull()
         try {
             val snapshot =
                 databaseReference
@@ -56,14 +43,6 @@ class FirebaseRepositoryImpl @Inject constructor(
             Log.e("RemoveMovie", "Error: ${e.message}")
         }
     }
-
-//    override suspend fun getMovie(): List<FilmDomain> {
-//        val snapshot = databaseReference.child("list_movies").get().await()
-//        return snapshot.children.mapNotNull { childSnapshot ->
-//            childSnapshot.getValue(FilmData::class.java)
-//        }
-//            .map { it.toDomain() } // Преобразуем FilmData в FilmDomain
-//    }
 
     override suspend fun getMovie(): List<SelectedMovie> {
         val snapshot = databaseReference.child("list_movies").get().await()
