--- app/src/main/java/com/pozmaxpav/cinemaopinion/domain/models/modelfirebase/FilmDomain.kt
@@ -1,5 +1,0 @@
-package com.pozmaxpav.cinemaopinion.domain.models.modelfirebase
-
-data class FilmDomain(
-    val titleFilm: String
-)
